Group page for unjoined group (#75)
* Now we can click every groups in all groups page. Also added white box for every group.

* Added join-leave button for activities instead of checkbox.
isMember will check if the user is a member of a specific group, group page has two views right now. One for members, the other is for non-members.

* Navigation from AllGroupsPage to any group page.
isMember is now working.
If the user is a member, then will be able to see everything about the group. If the user is not a member then will be able to see another view.

* Added Join group button for unjoined.
Changes according to Figma-prototype.

* Added checkered background.

* Changed SizedBox's height and font weight.

* Changes according to feedback from GitHub.